Created IndexedCollection.
- got FilteredCollection to work with it

The events fired by IC needs work, right now it fires

Index works by mapping all attributes x values as indexArr[attribute][value]
The downside of this is we don't get multi key indexes.

Events fired:

index:add/remove when a model is added to an index
index:created/emptied when a new

This means when adding one model multiple index:add will be fired, one
for each attribute. This makes it hard to use it with filters that act
on multiple attributes.

The big upside of this approach is realized in
FilteredCollection._update(), that simply finds the models to include by
intersecting the indexes for each attribute.

Other problems I ran into with current approach:

  - how to implement another filter for searching
  - how to limit the number of models visible